Update and rename GuessNumber.java to GuessingGame.java
